# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "description": "Retrieves all users. Admin only.",
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for email, fullnameTh, fullnameEn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., email:asc, createdAt:desc)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter criteria in JSON format or as query parameters",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all users retrieved successfully"
          },
          "403": {
            "description": "Forbidden - requires admin role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Creates a new user profile for authenticated users.",
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User profile created successfully"
          },
          "400": {
            "description": "Bad request - invalid input"
          },
          "401": {
            "description": "Unauthorized - requires authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "description": "Retrieves the profile of the currently authenticated user.",
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized - requires authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retrieves user by ID. Users can only access their own profile.",
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "403": {
            "description": "Forbidden - cannot access other users' profiles"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Updates a user by ID. Users can only update their own profile.",
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "403": {
            "description": "Forbidden - cannot update other users's profiles"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/email/{email}": {
      "get": {
        "description": "Retrieves user by email for any authenticated user.",
        "operationId": "UsersController_getUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "description": "User email address",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "401": {
            "description": "Unauthorized - requires authentication"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by email",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Updates a user by email. Admin only.",
        "operationId": "UsersController_updateUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "description": "User email address",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "403": {
            "description": "Forbidden - requires admin role"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user by email",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/company/user/{userId}": {
      "get": {
        "operationId": "CompanyController_getByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "The user ID of the company",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the company with the specified user ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get company by user ID",
        "tags": [
          "Company"
        ]
      }
    },
...
    "/freelance/user/{userId}": {
      "get": {
        "operationId": "FreelanceController_getByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "The user ID of the freelance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the freelance with the specified user ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Freelance"
                }
              }
            }
          },
          "404": {
            "description": "Freelance not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get freelance by user ID",
        "tags": [
          "Freelance"
        ]
      }
    },
...
  },
  "info": {
    "title": "DEPA API",
    "description": "The DEPA API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "fullnameTh": {
            "type": "string",
            "description": "User full name in Thai"
          },
          "fullnameEn": {
            "type": "string",
            "description": "User full name in English"
          },
          "about": {
            "type": "string",
            "description": "About the user"
          },
          "phoneNumber": {
            "type": "string",
            "description": "User phone number"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "website": {
            "type": "string",
            "description": "User website"
          },
          "location": {
            "type": "string",
            "description": "User location"
          },
          "industries": {
            "description": "Industries associated with the user",
            "example": [
              "Technology",
              "Healthcare"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "description": "User tags",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            },
            "example": [
              {
                "category": "skill",
                "name": "JavaScript"
              },
              {
                "category": "interest",
                "name": "AI"
              }
            ]
          },
          "channels": {
            "type": "array",
            "description": "User communication channels",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            },
            "example": [
              {
                "category": "email",
                "name": "support@example.com"
              },
              {
                "category": "LINE",
                "name": "@username"
              }
            ]
          },
          "specialists": {
            "type": "array",
            "description": "User specializations",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            },
            "example": [
              {
                "category": "programming",
                "name": "Backend Development"
              },
              {
                "category": "design",
                "name": "UX/UI"
              }
            ]
          }
        },
        "required": [
          "fullnameTh",
          "email",
          "industries",
          "channels",
          "specialists"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}